feat: Enhance bottom navigation with admin features
Introduces an admin view to the bottom navigation bar, enabling administrative actions and displaying admin-specific navigation items. This change also includes the addition of the 'Blog' link to the main navigation items in the header, ensuring consistent access across components. The "Top Brand" button in the header is now conditionally displayed, hidden when the admin is authenticated.

New icons for 'Dashboard' and 'Logout' have been added to support the admin functionality. The 'Hero' section now correctly labels the button as "Lihat Semua Blog" instead of "Lihat Semua Artikel".